# docs/openapi.yml
components:
  schemas:
...
    EsiScope:
      description: Enumeration of the ESI scopes
      enum:
      - esi-alliances.read_contacts.v1
      - esi-assets.read_assets.v1
      - esi-assets.read_corporation_assets.v1
      - esi-bookmarks.read_character_bookmarks.v1
      - esi-bookmarks.read_corporation_bookmarks.v1
      - esi-calendar.read_calendar_events.v1
      - esi-calendar.respond_calendar_events.v1
      - esi-characters.read_agents_research.v1
      - esi-characters.read_blueprints.v1
      - esi-characters.read_chat_channels.v1
      - esi-characters.read_contacts.v1
      - esi-characters.read_corporation_roles.v1
      - esi-characters.read_fatigue.v1
      - esi-characters.read_fw_stats.v1
      - esi-characters.read_loyalty.v1
      - esi-characters.read_medals.v1
      - esi-characters.read_notifications.v1
      - esi-characters.read_opportunities.v1
      - esi-characters.read_standings.v1
      - esi-characters.read_titles.v1
      - esi-characters.write_contacts.v1
      - esi-characterstats.read.v1
      - esi-clones.read_clones.v1
      - esi-clones.read_implants.v1
      - esi-contracts.read_character_contracts.v1
      - esi-contracts.read_corporation_contracts.v1
      - esi-corporations.read_blueprints.v1
      - esi-corporations.read_contacts.v1
      - esi-corporations.read_container_logs.v1
      - esi-corporations.read_corporation_membership.v1
      - esi-corporations.read_divisions.v1
      - esi-corporations.read_facilities.v1
      - esi-corporations.read_fw_stats.v1
      - esi-corporations.read_medals.v1
      - esi-corporations.read_standings.v1
      - esi-corporations.read_starbases.v1
      - esi-corporations.read_structures.v1
      - esi-corporations.read_titles.v1
      - esi-corporations.track_members.v1
      - esi-fittings.read_fittings.v1
      - esi-fittings.write_fittings.v1
      - esi-fleets.read_fleet.v1
      - esi-fleets.write_fleet.v1
      - esi-industry.read_character_jobs.v1
      - esi-industry.read_character_mining.v1
      - esi-industry.read_corporation_jobs.v1
      - esi-industry.read_corporation_mining.v1
      - esi-killmails.read_corporation_killmails.v1
      - esi-killmails.read_killmails.v1
      - esi-location.read_location.v1
      - esi-location.read_online.v1
      - esi-location.read_ship_type.v1
      - esi-mail.organize_mail.v1
      - esi-mail.read_mail.v1
      - esi-mail.send_mail.v1
      - esi-markets.read_character_orders.v1
      - esi-markets.read_corporation_orders.v1
      - esi-markets.structure_markets.v1
      - esi-planets.manage_planets.v1
      - esi-planets.read_customs_offices.v1
      - esi-search.search_structures.v1
      - esi-skills.read_skillqueue.v1
      - esi-skills.read_skills.v1
      - esi-ui.open_window.v1
      - esi-ui.write_waypoint.v1
      - esi-universe.read_structures.v1
      - esi-wallet.read_character_wallet.v1
      - esi-wallet.read_corporation_wallet.v1
      - esi-wallet.read_corporation_wallets.v1
      - publicData
      title: EsiScope
      type: string
...
    GetTokenOut:
      description: Model for ``GET /token`` responses.
      properties:
        callback:
          format: uri
          maxLength: 65536
          minLength: 1
          title: Callback
          type: string
        comments:
          title: Comments
          type: string
        created_on:
          format: date-time
          title: Created On
          type: string
        expires_on:
          format: date-time
          title: Expires On
          type: string
        owner:
          $ref: '#/components/schemas/GetUserOut'
        parent:
          title: Parent
          type: string
        token_type:
          $ref: '#/components/schemas/TokenType'
        uuid:
          title: Uuid
          type: string
      required:
      - created_on
      - owner
      - token_type
      - uuid
      title: GetTokenOut
      type: object
...
    GetUserOut:
      description: Model for ``GET /user/{character_name}`` responses
      properties:
        alliance:
          title: Alliance
          type: integer
        authorized_to_login:
          title: Authorized To Login
          type: boolean
        available_esi_scopes:
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Available Esi Scopes
          type: array
        character_id:
          title: Character Id
          type: integer
        character_name:
          title: Character Name
          type: string
        clearance_level:
          title: Clearance Level
          type: integer
        coalitions:
          items:
            type: string
          title: Coalitions
          type: array
        corporation:
          title: Corporation
          type: string
        created_on:
          format: date-time
          title: Created On
          type: string
        cumulated_mandatory_esi_scopes:
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Cumulated Mandatory Esi Scopes
          type: array
        is_ceo_of_alliance:
          title: Is Ceo Of Alliance
          type: boolean
        is_ceo_of_corporation:
          title: Is Ceo Of Corporation
          type: boolean
        tickered_name:
          title: Tickered Name
          type: string
        updated_on:
          format: date-time
          title: Updated On
          type: string
      required:
      - available_esi_scopes
      - character_id
      - character_name
      - clearance_level
      - coalitions
      - created_on
      - cumulated_mandatory_esi_scopes
      - is_ceo_of_alliance
      - is_ceo_of_corporation
      - tickered_name
      - updated_on
      title: GetUserOut
      type: object
    GetUserShortOut:
      description: Model for an element of ``GET /user`` response
      properties:
        character_id:
          title: Character Id
          type: integer
        character_name:
          title: Character Name
          type: string
      required:
      - character_id
      - character_name
      title: GetUserShortOut
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PostTokenDynIn:
      description: Model for ``POST /token/dyn`` requests.
      properties:
        callback:
          format: uri
          maxLength: 65536
          minLength: 1
          title: Callback
          type: string
        comments:
          title: Comments
          type: string
      required:
      - callback
      title: PostTokenDynIn
      type: object
    PostTokenDynOut:
      description: Model for ``POST /token/dyn`` reponses.
      properties:
        tkn:
          title: Tkn
          type: string
      required:
      - tkn
      title: PostTokenDynOut
      type: object
    PostTokenPerIn:
      description: Model for ``POST /token/per`` requests.
      properties:
        callback:
          format: uri
          maxLength: 65536
          minLength: 1
          title: Callback
          type: string
        comments:
          title: Comments
          type: string
      required:
      - callback
      title: PostTokenPerIn
      type: object
    PostTokenPerOut:
      description: Model for ``POST /token/per`` reponses.
      properties:
        tkn:
          title: Tkn
          type: string
      required:
      - tkn
      title: PostTokenPerOut
      type: object
    PostTokenUseFromDynIn:
      description: Model for ``POST /token/use/from/dyn`` requests.
      properties:
        scopes:
          default:
          - publicData
          items:
            $ref: '#/components/schemas/EsiScope'
          title: Scopes
          type: array
        state_code:
          title: State Code
          type: string
      title: PostTokenUseFromDynIn
      type: object
    PostTokenUseFromDynOut:
      description: Model for ``POST /token/use/from/dyn`` reponses.
      properties:
        login_url:
          title: Login Url
          type: string
        state_code:
          title: State Code
          type: string
      required:
      - login_url
      - state_code
      title: PostTokenUseFromDynOut
      type: object
    PostUseFromPerOut:
      description: Model for ``POST /token/use/from/per`` reponses.
      properties:
        user_token:
          title: User Token
          type: string
      required:
      - user_token
      title: PostUseFromPerOut
      type: object
...
    TokenType:
      description: Enumeration containing the various token types.
      enum:
      - dyn
      - per
      - use
      title: TokenType
      type: string
    ValidationError:
      properties:
        loc:
          items:
            type: string
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    sni__api__routers__teamspeak__PostAuthStartOut:
      description: Model for `POST /teamspeak/auth/start` responses.
      properties:
        challenge_nickname:
          title: Challenge Nickname
          type: string
        expiration_datetime:
          format: date-time
          title: Expiration Datetime
          type: string
        user:
          $ref: '#/components/schemas/GetUserShortOut'
      required:
      - expiration_datetime
      - challenge_nickname
      - user
      title: PostAuthStartOut
      type: object
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.0.2
paths:
...
  /teamspeak/auth/complete:
    post:
      description: 'Completes an authentication challenge for the owner of the token.
        See the

        `POST /teamspeak/auth/start` documentation.'
      operationId: post_auth_complete_teamspeak_auth_complete_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Completes a teamspeak authentication challenge
      tags:
      - Teamspeak
  /teamspeak/auth/start:
    post:
      description: 'Starts a new authentication challenge for the owner of the token.
        A random

        nickname is returned (see `PostAuthStartOut` for more details), and the

        user has 1 minute to update its teamspeak nickname to it, and then call

        `POST /teamspeak/auth/complete`.'
      operationId: port_auth_start_teamspeak_auth_start_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sni__api__routers__teamspeak__PostAuthStartOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Starts a teamspeak authentication challenge
      tags:
      - Teamspeak
  /token:
    delete:
      description: Manually deletes a token (of any type). Requires a clearance level
        of 10.
      operationId: delete_token_token_delete
      parameters:
      - in: query
        name: uuid
        required: true
        schema:
          format: uuid4
          title: Uuid
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Manually delete a token
      tags:
      - Authentication & tokens
    get:
      description: 'Returns informations about the token currently being used. Requires
        a

        clearance level of 0 or more.'
      operationId: get_token_token_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTokenOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get basic informations about the current token
      tags:
      - Authentication & tokens
  /token/dyn:
    post:
      description: 'Creates a new dynamic app token. Must be called with a permanent
        app token,

        and the owner must have a clearance level of 10.'
      operationId: post_token_dyn_token_dyn_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTokenDynIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostTokenDynOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new dynamic app token
      tags:
      - Authentication & tokens
  /token/per:
    post:
      description: 'Creates a new permanent app token. Must be called with a permanent
        app

        token, and the owner must have a clearance level of 10.'
      operationId: post_token_per_token_per_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTokenPerIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostTokenPerOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new permanent app token
      tags:
      - Authentication & tokens
  /token/use/from/dyn:
    post:
      description: 'Authenticates an application dynamic token and returns a `state
        code` and

        an URL at which the user can authenticate to the EVE SSO. Once that is

        done, SNI issues a GET request to the app predefined callback, with that

        state code and the user token. Requires the owner of the token to have a

        clearance level of 0 or more.'
      operationId: post_token_use_from_dyn_token_use_from_dyn_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTokenUseFromDynIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostTokenUseFromDynOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new user token with a dynamic app token
      tags:
      - Authentication & tokens
  /token/use/from/per:
    post:
      description: 'Authenticates an application permanent token and returns a user
        token tied

        to the owner of that app token.'
      operationId: post_token_use_from_per_token_use_from_per_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostUseFromPerOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new user token with a permanent app token
      tags:
      - Authentication & tokens
...